Enable providers to provide authentication parameters
Azure AD expects a resource to authenticate against (this is different
that a scope) - this is sent as a `resource`parameter to the oauth2
flow. It would be possible for users to specify `{ providerParams: {
resource: “foo” } }` but I imagine this would be a common “gotcha”.

Instead enable provider to add additional parameters to the request via
`provider.providerParams`. These are merged into the query object, and
existing settings.providerParams will take precedence in the merge.